requirements.txt
flask
gunicorn
psycopg2-binary
requests
telethon
gunicorn
google-generativeai
groq
tweepy
PySocks
